Move Pkcs12Builder and all required types to Common
