+++ src/libraries/Common/src/System/Security/Cryptography/CryptographicAttributeObject.cs
@@ -7,7 +7,13 @@ using Internal.Cryptography;
 
 namespace System.Security.Cryptography
 {
-    public sealed class CryptographicAttributeObject
+#if BUILDING_PKCS
+    public
+#else
+    #pragma warning disable CA1510, CA1512
+    internal
+#endif
+    sealed class CryptographicAttributeObject
     {
         //
         // Constructors.
