+++ src/libraries/Common/src/System/Security/Cryptography/Pkcs/Pkcs12SafeContents.cs
@@ -4,7 +4,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Formats.Asn1;
-using System.Linq;
 using System.Security.Cryptography.Asn1.Pkcs12;
 using System.Security.Cryptography.Asn1.Pkcs7;
 using System.Security.Cryptography.X509Certificates;
@@ -12,7 +11,13 @@ using Internal.Cryptography;
 
 namespace System.Security.Cryptography.Pkcs
 {
-    public sealed class Pkcs12SafeContents
+#if BUILDING_PKCS
+    public
+#else
+    #pragma warning disable CA1510, CA1512
+    internal
+#endif
+    sealed class Pkcs12SafeContents
     {
         private ReadOnlyMemory<byte> _encrypted;
         private List<Pkcs12SafeBag>? _bags;
@@ -287,7 +292,7 @@ namespace System.Security.Cryptography.Pkcs
 
             if (_bags == null)
             {
-                return Enumerable.Empty<Pkcs12SafeBag>();
+                return [];
             }
 
             return _bags.AsReadOnly();
@@ -359,7 +364,7 @@ namespace System.Security.Cryptography.Pkcs
 
                 bag ??= new Pkcs12SafeBag.UnknownBag(serializedBags[i].BagId, bagValue);
 
-                bag.Attributes = SignerInfo.MakeAttributeCollection(serializedBags[i].BagAttributes);
+                bag.Attributes = PkcsHelpers.MakeAttributeCollection(serializedBags[i].BagAttributes);
                 bags.Add(bag);
             }
