+++ src/libraries/Common/src/System/Security/Cryptography/Pkcs/Pkcs9LocalKeyId.cs
@@ -7,7 +7,13 @@ using Internal.Cryptography;
 
 namespace System.Security.Cryptography.Pkcs
 {
-    public sealed class Pkcs9LocalKeyId : Pkcs9AttributeObject
+#if BUILDING_PKCS
+    public
+#else
+    #pragma warning disable CA1510, CA1512
+    internal
+#endif
+    sealed class Pkcs9LocalKeyId : Pkcs9AttributeObject
     {
         private byte[]? _lazyKeyId;
