+++ src/libraries/System.Security.Cryptography.Pkcs/src/Internal/Cryptography/Pal/AnyOS/ManagedPal.Decode.cs
@@ -50,7 +50,7 @@ namespace Internal.Cryptography.Pal.AnyOS
                 }
             }
 
-            unprotectedAttributes = SignerInfo.MakeAttributeCollection(data.UnprotectedAttributes);
+            unprotectedAttributes = PkcsHelpers.MakeAttributeCollection(data.UnprotectedAttributes);
 
             var recipientInfos = new List<RecipientInfo>();
